+++ README.md
@@ -2,18 +2,36 @@
 
 Golang + Mongodb + react-admin
 
-# Dev
+## Docker
+```sh
+#通过 docker 启动项目 
+docker-compose up -d
+
+# 查看特定 container （如：api）日志
+docker-compose logs api
+
+# 访问本地 127.0.0.1:3000 如数据、操作等均无误则说明启动成功
+```
+
+## Dev
 ```sh
 # /src/config.yml 配置好 mongodb connection
 # api /src
 cd src
 go run .
-# app /src/app
+
+# app /frontend
 # yarn 不行就试试 npm
-cd app
+cd ../frontend
 yarn & yarn start
 ```
-## Mitre ATT&CK 编辑
+
+```sh
+# 将 mongodb 中string字段转化为 Date，方便排序
+> db.mitre_attack.find().forEach(function(doc){doc.created = new Date(doc.created);db.mitre_attack.save(doc)});
+> db.mitre_attack.find().forEach(function(doc){doc.modified = new Date(doc.modified);db.mitre_attack.save(doc)});
+```
+### Mitre ATT&CK 编辑
 
 ![](./attackPatterns.png)
