+++ database/init-mongo.js
@@ -1,0 +1,12 @@
+db.createUser(
+    {
+        user: "attackSeaman",
+        pwd: "cuccs1sgreat",
+        roles: [
+            {
+                role: "readWrite",
+                db: "attackSeaman"
+            }
+        ]
+    }
+);
