+++ frontend/Dockerfile
@@ -1,0 +1,23 @@
+# Pull latest official node image
+FROM node:latest
+
+# Expose ports
+EXPOSE 3000
+EXPOSE 35729
+
+# Set working directory
+WORKDIR /app
+
+# Add /app/node_modules/.bin to environment variables
+ENV PATH /app/node_modules/.bin:$PATH
+
+# Copy package files and install app dependencies
+COPY package.json /app/package.json
+COPY package-lock.json /app/package-lock.json
+RUN npm install
+
+# Add React app to working directory
+ADD . /app
+
+# Start the React app
+CMD ["npm", "start"]
