+++ frontend/src/MyUrlField.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { withStyles } from '@material-ui/core/styles';
+import LaunchIcon from '@material-ui/icons/Launch';
+
+const styles = {
+    link: {
+        textDecoration: 'none',
+    },
+    icon: {
+        width: '0.5em',
+        paddingLeft: 2,
+    },
+};
+
+const MyUrlField = ({ record = {}, source, classes }:any) =>
+    <a href={record[source]} className={classes.link}>
+        {record[source]}
+        <LaunchIcon className={classes.icon} />
+    </a>;
+
+export default withStyles(styles)(MyUrlField);
