+++ frontend/src/react-app-env.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react-scripts" />
+declare module "ra-data-json-server"
+declare module "react-admin"
