+++ src/config.yml
@@ -12,6 +12,6 @@ server:
     Access-Control-Expose-Headers: "X-Total-Count"
 
 database: # for database see (configure database section)
-  dbname: qi
-  connection: mongodb://localhost:27017
+  dbname: attackSeaman
+  connection: mongodb://attackSeaman:cuccs1sgreat@mongodb:27017
