+++ src/go.mod
@@ -3,19 +3,13 @@ module github.com/zGina/Attack-Seaman/src
 go 1.12
 
 require (
-	github.com/gin-contrib/sse v0.0.0-20190301062529-5545eab6dad3 // indirect
+	github.com/gin-contrib/sse v0.1.0 // indirect
 	github.com/gin-gonic/gin v1.3.0
-	github.com/go-stack/stack v1.8.0 // indirect
-	github.com/golang/protobuf v1.3.1 // indirect
-	github.com/golang/snappy v0.0.1 // indirect
+	github.com/golang/protobuf v1.4.3 // indirect
+	github.com/gotify/configor v1.0.2
 	github.com/gotify/server v1.2.1
-	github.com/lotteryjs/configor v1.0.2
-	github.com/mattn/go-isatty v0.0.7 // indirect
-	github.com/stretchr/testify v1.6.1
-	github.com/ugorji/go/codec v0.0.0-20190316192920-e2bddce071ad // indirect
-	github.com/xdg/scram v0.0.0-20180814205039-7eeb5667e42c // indirect
-	github.com/xdg/stringprep v1.0.0 // indirect
+	github.com/mattn/go-isatty v0.0.12 // indirect
+	github.com/ugorji/go v1.2.4 // indirect
 	go.mongodb.org/mongo-driver v1.4.6
-	golang.org/x/crypto v0.0.0-20190530122614-20be4c3c3ed5
 	gopkg.in/go-playground/validator.v8 v8.18.2
 )
