--- src/test.http
@@ -1,29 +1,0 @@
-
-### typicode users
-GET http://jsonplaceholder.typicode.com/users?_end=5&_order=DESC&_sort=id&_start=0
-
-
-### local 分页获取用户列表
-GET http://localhost:6868/users?_end=5&_order=DESC&_sort=id&_start=0
-
-### local 获取用户，ID 不正确
-GET http://localhost:6868/users?id=11
-
-### local 获取用户，ID 正确
-GET http://localhost:6868/users?id=5c933ae7a49cac27417def6f&id=5c933ae7a49cac27417def70
-
-### local 删除用户，ID 正确
-DELETE http://localhost:6868/users/5c99bd941ba7b2304ad8c52b
-
-### local 分页获取文章列表
-GET http://localhost:6868/posts?_end=5&_order=DESC&_sort=id&_start=0&userId=5c938131ca447e20e7b66974
-
-### local 获取文章，ID 不正确
-GET http://localhost:6868/posts/11
-
-### local 获取文章，ID 正确
-GET http://localhost:6868/posts/5c92e6199929adef73bceea1
-
-### local 删除文章，ID 正确
-DELETE http://localhost:6868/posts/5c98678fbf0b9c5d8699e587
-
